@@ -4,16 +4,16 @@
   "homepage": "http://jesusmontano.github.io/randomizer",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.10.2",
+    "@mui/material": "^5.4.2",
+    "@mui/styles": "^5.4.2",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "onsenui": "^2.10.10",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-onsenui": "^1.11.2",
-    "react-scripts": "3.4.1",
-    "react-tradeshift-ui": "^2.6.0"
+    "react-scripts": "3.4.1"
   },
   "scripts": {
     "start": "react-scripts start",
